renamed 'forms/create.py into 'forms/common.py'
